Module darkfi::blockchain

Re-exports

pub use block_store::Block;

pub use block_store::BlockInfo;

pub use block_store::BlockOrderStore;

pub use block_store::BlockOrderStoreOverlay;

pub use block_store::BlockStore;

pub use block_store::BlockStoreOverlay;

pub use header_store::Header;

pub use header_store::HeaderStore;

pub use header_store::HeaderStoreOverlay;

pub use slot_store::validate_slot;

pub use slot_store::SlotStore;

pub use slot_store::SlotStoreOverlay;

pub use tx_store::PendingTxOrderStore;

pub use tx_store::PendingTxStore;

pub use tx_store::TxStore;

pub use tx_store::TxStoreOverlay;

pub use contract_store::ContractStateStore;

pub use contract_store::ContractStateStoreOverlay;

pub use contract_store::WasmStore;

pub use contract_store::WasmStoreOverlay;

Modules

block_store

Block related definitions and storage implementations

contract_store

Contracts and Wasm storage implementations

header_store

Header definition and storage implementation

slot_store

Slots storage implementation

tx_store

Transactions related storage implementations

Structs

Blockchain

Structure holding all sled trees that define the concept of Blockchain.

BlockchainOverlay

Overlay structure over a Blockchain instance.

Functions

parse_record

Parse a sled record in the form of a tuple (key, value).

parse_u64_key_record

Parse a sled record with a u64 keyin the form of a tuple (key, value).

Type Definitions

BlockchainOverlayPtr

Atomic pointer to blockchain overlay.

SledDbOverlayPtr

Atomic pointer to sled db overlay.