Module darkfi::blockchain::block_store

Block related definitions and storage implementations

Structs

Block

This struct represents a tuple of the form (magic, header, txs, producer, slots). The header and transactions are stored as hashes, while slots are stored as integers, serving as pointers to the actual data in the sled database.

BlockInfo

Structure representing full block data.

BlockOrder

Auxiliary structure used to keep track of blocks order.

BlockOrderStore

The BlockOrderStore is a sled tree storing the order of the blockchain’s blocks, where the key is the order number, and the value is the blocks’ hash. BlockStore can be queried with this hash.

BlockOrderStoreOverlay

Overlay structure over a BlockOrderStore instance.

BlockProducer

This struct represents Block producer information.

BlockStore

The BlockStore is a sled tree storing all the blockchain’s blocks where the key is the blocks’ hash, and value is the serialized block.

BlockStoreOverlay

Overlay structure over a BlockStore instance.

Constants

BLOCK_MAGIC_BYTES 🔒

Block magic bytes

BLOCK_VERSION

Block version number

SLED_BLOCK_ORDER_TREE 🔒

BlockOrder sled tree

SLED_BLOCK_TREE 🔒

Block sled tree