darkfi::blockchain
Module contract_store

Contracts and Wasm storage implementations

Structs

ContractStore

The ContractStore is a structure representing all sled trees related to storing the blockchain’s contracts information.

ContractStoreOverlay

Overlay structure over a ContractStore instance.

Constants

SLED_BINCODE_TREE

SLED_CONTRACTS_TREE

SMART_CONTRACT_ZKAS_DB_NAME

The hardcoded db name for the zkas circuits database tree