darkfi::blockchain

Module header_store

Header definition and storage implementation

Structs

Header

This struct represents a tuple of the form (version, previous, height, timestamp, nonce, merkle_tree).

HeaderHash

HeaderStore

The HeaderStore is a structure representing all sled trees related to storing the blockchain’s blocks’s header information.

HeaderStoreOverlay

Overlay structure over a HeaderStore instance.

Enums

PowData

Struct representing the Proof of Work used in a block.

Constants

SLED_HEADER_TREE

SLED_SYNC_HEADER_TREE