darkfi::blockchain

Module monero

Monero definitions needed for merge mining

Re-exports

pub use merkle_tree_parameters::MerkleTreeParameters;

Modules

fixed_array

keccak

merkle_proof

merkle_tree_parameters

utils

Structs

MoneroPowData

This struct represents all the Proof of Work information required for merge mining.

Functions

create_ordered_tx_hashes_from_block

Create a set of ordered transaction hashes from a Monero block

extract_aux_merkle_root

Extract the Monero block hash from the coinbase transaction’s extra field

extract_aux_merkle_root_from_block

Extract the Monero block hash from the coinbase transaction’s extra field

insert_aux_chain_mr_and_info_into_block

Inserts aux chain merkle root and info into a Monero block

monero_block_deserialize

Try to decode a monero::Block given a hex blob

parse_extra_field_truncate_on_error 🔒

Parsing an extra field from bytes will always return an extra field with subfields that could be read even if it does not represent the original As per Monero consensus rules, an error here will not represent failure to deserialize a block, so no need to error here.

Type Aliases

AuxChainHashes