darkfi::blockchain::monero::fixed_array

Struct MaxSizeVec

pub struct MaxSizeVec<T, const MAX_SIZE: usize>
where
    T: Send + Sync,
{
    vec: Vec<T>,
    _marker: PhantomData<T>,
}

A vector that has a maximum size of MAX_SIZE

Fields
vec: Vec<T>_marker: PhantomData<T>
Implementations

impl<T, const MAX_SIZE: usize> MaxSizeVec<T, MAX_SIZE>
where T: Send + Sync,

pub fn new() -> Self

Creates a new MaxSizeVec with a capacity of MAX_SIZE

pub fn new_with_data(data: Vec<T>) -> Result<Self>

Creates a new MaxSizeVec with the given data. Returns an error if the data length exceeds MAX_SIZE.

pub fn from_items_truncate(items: Vec<T>) -> Self

Creates a MaxSizeVec from the given items, truncating if needed

pub fn into_vec(self) -> Vec<T>

Consumes MaxSizeVec and returns the inner Vec<T>

pub fn max_size(&self) -> usize

Returns the maximum size of the MaxSizeVec

pub fn push(&mut self, item: T) -> Result<()>

Pushes an item to the MaxSizeVec