Struct darkfi::blockchain::BlockchainOverlay

pub struct BlockchainOverlay {
    pub overlay: SledDbOverlayPtr,
    pub headers: HeaderStoreOverlay,
    pub blocks: BlockStoreOverlay,
    pub transactions: TxStoreOverlay,
    pub contracts: ContractStoreOverlay,
}

Overlay structure over a Blockchain instance.

Fields
overlay: SledDbOverlayPtr

Main [sled_overlay::SledDbOverlay] to the sled db connection

headers: HeaderStoreOverlay

Headers overlay

blocks: BlockStoreOverlay

Blocks overlay

transactions: TxStoreOverlay

Transactions overlay

contracts: ContractStoreOverlay

Contract overlay

Implementations

impl BlockchainOverlay

pub fn new(blockchain: &Blockchain) -> Result<BlockchainOverlayPtr>

Instantiate a new BlockchainOverlay over the given Blockchain instance.

pub fn is_empty(&self) -> Result<bool>

Check if blockchain contains any blocks

pub fn last(&self) -> Result<(u32, HeaderHash)>

Retrieve the last block height and hash.

pub fn last_block(&self) -> Result<BlockInfo>

Retrieve the last block info.

pub fn last_block_height(&self) -> Result<u32>

Retrieve the last block height.

pub fn last_block_timestamp(&self) -> Result<Timestamp>

Retrieve the last block timestamp.

pub fn add_block(&self, block: &BlockInfo) -> Result<HeaderHash>

Insert a given BlockInfo into the overlay. This functions wraps all the logic of separating the block into specific data that can be fed into the different trees of the overlay. Upon success, the functions returns the block hash that were given and appended to the overlay. Since we are adding to the overlay, we don’t need to exeucte the writes atomically.

pub fn has_block(&self, block: &BlockInfo) -> Result<bool>

Check if the given BlockInfo is in the database and all trees.

pub fn get_blocks_by_hash( &self, hashes: &[HeaderHash], ) -> Result<Vec<BlockInfo>>

Retrieve BlockInfos by given hashes. Fails if any of them is not found.

fn get_blocks_infos(&self, blocks: &[Block]) -> Result<Vec<BlockInfo>>

Retrieve all BlockInfo for given slice of Block. Fails if any of them is not found

pub fn get_blocks_txs_hashes( &self, hashes: &[HeaderHash], ) -> Result<Vec<TransactionHash>>

Retrieve Blocks by given hashes and return their transactions hashes.

pub fn checkpoint(&self)

Checkpoint overlay so we can revert to it, if needed.

pub fn revert_to_checkpoint(&self) -> Result<()>

Revert to current overlay checkpoint.

pub fn full_clone(&self) -> Result<BlockchainOverlayPtr>

Auxiliary function to create a full clone using SledDbOverlay::clone, generating new pointers for the underlying overlays.