Module darkfi::consensus::task

async tasks to utilize the protocols

Modules

block_sync 🔒

consensus_sync 🔒

proposal 🔒

Functions

block_sync_task

async task used for block syncing.

consensus_sync_task

async task used for consensus state syncing. Returns flag if node is not connected to other peers or consensus hasn’t started, so it can immediately start proposing proposals.

proposal_task

async task used for participating in the consensus protocol