darkfi::error

Enum Error

pub enum Error {

    ParseFailed(&'static str),
    ParseIntError(ParseIntError),
    ParseFloatError(ParseFloatError),
    UrlParseError(ParseError),
    UrlParse(String),
    AddrParseError(AddrParseError),
    TokenParseError,
    TryFromSliceError(TryFromSliceError),
    SemverError(String),
    DecodeError(&'static str),
    EncodeError(&'static str),
    NonMinimalVarInt,
    Utf8Error(FromUtf8Error),
    StrUtf8Error(Utf8Error),
    JsonParseError(String),
    JsonGenerateError(String),
    TomlDeserializeError(Error),
    Bs58DecodeError(Error),
    HexDecodeError(FromHexError),
    BadOperationType,
    InvalidDialerScheme,
    InvalidListenerScheme,
    UnsupportedTransport(String),
    UnsupportedTransportUpgrade(String),
    InvalidTransportRequest,
    ConnectFailed(String),
    TimeoutError(TimeoutError),
    ConnectTimeout,
    ChannelStopped,
    ChannelTimeout,
    SeedFailed,
    NetworkServiceStopped,
    BindFailed(String),
    AcceptConnectionFailed(String),
    AcceptTlsConnectionFailed(String),
    ConnectorStopped(String),
    NetworkOperationFailed,
    MissingDispatcher,
    MessageInvalid,
    MeteringLimitExceeded,
    ArtiError(Error),
    MalformedPacket,
    DecodePacket(String),
    SocksError(String),
    NoSocks5UrlFound,
    NoUrlFound,
    TorError(String),
    NetworkNotConnected,
    P2PNetworkStopped,
    InvalidHostColor,
    HostDoesNotExist,
    HostStateBlocked(String, String),
    PlonkError(String),
    WrongWitnessType(usize),
    WrongWitnessesCount,
    WrongPublicInputsCount,
    NoteDecryptionFailed(String),
    KeypairPathNotFound,
    PublicKeyFromBytes,
    CoinFromBytes,
    SecretKeyFromBytes,
    PublicKeyFromStr,
    SecretKeyFromStr,
    InvalidAddress,
    TxRcptDecryptionError,
    Blake3FromHexError(HexError),
    UnsupportedChain,
    JsonRpcError((i32, String)),
    RpcServerError(RpcError),
    RpcConnectionsExhausted,
    RpcServerStopped,
    RpcClientStopped,
    UnexpectedJsonRpc(String),
    UnknownNodeError,
    InvalidPublicInputsError,
    InvalidSignature,
    StateTransitionError,
    ForksNotFound,
    ExtendedChainIndexNotFound,
    ProposalHashesMissmatchError,
    ProposalHeadersMissmatchError,
    TransferTxVerification,
    ErroneousTxsDetected,
    ProposalTaskStopped,
    ProposalAlreadyExists,
    ConsensusTaskStopped,
    MinerTaskStopped,
    GarbageCollectionTaskStopped,
    PoWTotalWorkIsZero,
    PoWCuttofCalculationError,
    PoWInvalidTimestamp,
    PoWInvalidOutHash,
    ContractsStatesRootNotFoundError,
    ContractsStatesRootError(String, String),
    MoneroHashingError(String),
    MoneroNumberOfChainZero,
    MoneroMergeMineError(String),
    RandomXError(String),
    DatabaseError(String),
    SledError(Error),
    SledTransactionError(TransactionError),
    TransactionNotFound(String),
    TransactionAlreadySeen,
    InvalidInputLengths,
    HeaderNotFound(String),
    BlockIsInvalid(String),
    BlockVersionIsInvalid(u8),
    BlockAlreadyExists(String),
    BlockNotFound(String),
    BlockHeightNotFound(u32),
    BlockDifficultyNotFound(u32),
    BlockStateInverseDiffNotFound(u32),
    BlockContainsNoTransactions(String),
    ContractNotFound(String),
    ContractStateNotFound,
    ContractAlreadyInitialized,
    ZkasBincodeNotFound,
    WasmerCompileError(String),
    WasmerExportError(String),
    WasmerRuntimeError(String),
    WasmerInstantiationError(String),
    WasmerMemoryError(String),
    WasmerOomError(String),
    ContractError(ContractError),
    DarkTreeError(DarkTreeError),
    WasmBincodeNotFound,
    ContractInitError(u64),
    ContractExecError(u64),
    WasmFunctionAclDenied,
    EventNotFound(String),
    EventIsInvalid,
    Io(ErrorKind),
    InfallibleError(String),
    AsyncChannelSendError(String),
    AsyncChannelRecvError(String),
    LogInitError(String),
    ValueIsNotObject,
    ConfigNotFound,
    ConfigInvalid,
    ConfigError(String),
    ZkasDecoderError(String),
    InvalidClock,
    UnsupportedOS,
    BackwardsTime(SystemTimeError),
    DetachedTaskStopped,
    AdditionOverflow,
    SubtractionUnderflow,
    ClientFailed(ClientFailed),
    TxVerifyFailed(TxVerifyFailed),
    ClockOutOfSync(String),
    GeodeNeedsGc,
    GeodeFileNotFound,
    GeodeChunkNotFound,
    GeodeFileRouteNotFound,
    GeodeChunkRouteNotFound,
    DagSyncFailed,
    MaliciousFlood,
    Custom(String),

}

General library errors used throughout the codebase.

Variants

ParseFailed(&'static str)

ParseIntError(ParseIntError)

ParseFloatError(ParseFloatError)

UrlParseError(ParseError)

UrlParse(String)

AddrParseError(AddrParseError)

TokenParseError

TryFromSliceError(TryFromSliceError)

SemverError(String)

DecodeError(&'static str)

EncodeError(&'static str)

NonMinimalVarInt

Utf8Error(FromUtf8Error)

StrUtf8Error(Utf8Error)

JsonParseError(String)

JsonGenerateError(String)

TomlDeserializeError(Error)

Bs58DecodeError(Error)

HexDecodeError(FromHexError)

BadOperationType

InvalidDialerScheme

InvalidListenerScheme

UnsupportedTransport(String)

UnsupportedTransportUpgrade(String)

InvalidTransportRequest

ConnectFailed(String)

TimeoutError(TimeoutError)

ConnectTimeout

ChannelStopped

ChannelTimeout

SeedFailed

NetworkServiceStopped

BindFailed(String)

AcceptConnectionFailed(String)

AcceptTlsConnectionFailed(String)

ConnectorStopped(String)

NetworkOperationFailed

MissingDispatcher

MessageInvalid

MeteringLimitExceeded

ArtiError(Error)

MalformedPacket

DecodePacket(String)

SocksError(String)

NoSocks5UrlFound

NoUrlFound

TorError(String)

NetworkNotConnected

P2PNetworkStopped

InvalidHostColor

HostDoesNotExist

HostStateBlocked(String, String)

PlonkError(String)

WrongWitnessType(usize)

WrongWitnessesCount

WrongPublicInputsCount

NoteDecryptionFailed(String)

KeypairPathNotFound

PublicKeyFromBytes

CoinFromBytes

SecretKeyFromBytes

PublicKeyFromStr

SecretKeyFromStr

InvalidAddress

TxRcptDecryptionError

Blake3FromHexError(HexError)

UnsupportedChain

JsonRpcError((i32, String))

RpcServerError(RpcError)

RpcConnectionsExhausted

RpcServerStopped

RpcClientStopped

UnexpectedJsonRpc(String)

UnknownNodeError

InvalidPublicInputsError

InvalidSignature

StateTransitionError

ForksNotFound

ExtendedChainIndexNotFound

ProposalHashesMissmatchError

ProposalHeadersMissmatchError

TransferTxVerification

ErroneousTxsDetected

ProposalTaskStopped

ProposalAlreadyExists

ConsensusTaskStopped

MinerTaskStopped

GarbageCollectionTaskStopped

PoWTotalWorkIsZero

PoWCuttofCalculationError

PoWInvalidTimestamp

PoWInvalidOutHash

ContractsStatesRootNotFoundError

ContractsStatesRootError(String, String)

MoneroHashingError(String)

MoneroNumberOfChainZero

MoneroMergeMineError(String)

RandomXError(String)

DatabaseError(String)

SledError(Error)

SledTransactionError(TransactionError)

TransactionNotFound(String)

TransactionAlreadySeen

InvalidInputLengths

HeaderNotFound(String)

BlockIsInvalid(String)

BlockVersionIsInvalid(u8)

BlockAlreadyExists(String)

BlockNotFound(String)

BlockHeightNotFound(u32)

BlockDifficultyNotFound(u32)

BlockStateInverseDiffNotFound(u32)

BlockContainsNoTransactions(String)

ContractNotFound(String)

ContractStateNotFound

ContractAlreadyInitialized

ZkasBincodeNotFound

WasmerCompileError(String)

WasmerExportError(String)

WasmerRuntimeError(String)

WasmerInstantiationError(String)

WasmerMemoryError(String)

WasmerOomError(String)

ContractError(ContractError)

DarkTreeError(DarkTreeError)

WasmBincodeNotFound

ContractInitError(u64)

ContractExecError(u64)

WasmFunctionAclDenied

EventNotFound(String)

EventIsInvalid

Io(ErrorKind)

InfallibleError(String)

AsyncChannelSendError(String)

AsyncChannelRecvError(String)

LogInitError(String)

ValueIsNotObject

ConfigNotFound

ConfigInvalid

ConfigError(String)

ZkasDecoderError(String)

InvalidClock

UnsupportedOS

BackwardsTime(SystemTimeError)

DetachedTaskStopped

AdditionOverflow

SubtractionUnderflow

ClientFailed(ClientFailed)

TxVerifyFailed(TxVerifyFailed)

ClockOutOfSync(String)

GeodeNeedsGc

GeodeFileNotFound

GeodeChunkNotFound

GeodeFileRouteNotFound

GeodeChunkRouteNotFound

DagSyncFailed

MaliciousFlood

Custom(String)

Implementations

impl Error

pub fn retrieve_erroneous_txs(&self) -> Result<Vec<Transaction>>

Auxiliary function to retrieve the vector of erroneous transactions from a TxVerifyFailed error. In any other case, we return the error itself.