darkfi::event_graph
Module proto

P2P protocol implementation for the Event Graph

Structs

EventPut

A P2P message representing publishing an event on the network

EventRep

A P2P message representing an event reply

EventReq

A P2P message representing an event request

MovingWindow 🔒

ProtocolEventGraph

P2P protocol implementation for the Event Graph.

TipRep

A P2P message representing a reply for the peer’s DAG tips

TipReq

A P2P message representing a request for a peer’s DAG tips

Constants

MALICIOUS_THRESHOLD 🔒

Malicious behaviour threshold. If the threshold is reached, we will drop the peer from our P2P connection.

RATELIMIT_EXPIRY_TIME 🔒

Rolling length of the window

RATELIMIT_MIN_COUNT 🔒

Ratelimit kicks in above this count

RATELIMIT_SAMPLE_IDX 🔒

Sample point used to calculate sleep time when ratelimit is active

RATELIMIT_SAMPLE_SLEEP 🔒

Sleep for this amount of time when count == RATE_LIMIT_SAMPLE_IDX.

WINDOW_EXPIRY_TIME 🔒

Rolling length of the window

WINDOW_MAXSIZE 🔒

Global limit of messages per window