darkfi::geode

Struct FileSequence

pub struct FileSequence {
    files: Vec<(PathBuf, u64)>,
    current_file: Option<File>,
    current_file_index: Option<usize>,
    position: u64,
    auto_set_len: bool,
}

FileSequence is an object that implements AsyncRead, AsyncSeek, and AsyncWrite for an ordered list of (file path, file size).

You can use it to read and write from/to a list a file, without having to manage individual file operations explicitly.

This allows seamless handling of multiple files as if they were a single continuous file. It automatically opens the next file in the list when the current file is exhausted.

It’s also made so that files in files that do not exist on the filesystem will get skipped, without returning an Error. All files you want to read, write, and seek to should be created before using the FileSequence.

Fields
files: Vec<(PathBuf, u64)>

List of (file path, file size). File sizes are not the sizes of the files as they currently are on the file system, but the sizes we want

current_file: Option<File>

Currently opened file

current_file_index: Option<usize>

Index of the currently opened file in the files vector

position: u64auto_set_len: bool

Set to true to automatically set the length of the file on the filesystem to it’s size as defined in the files vector, after a write

Implementations

impl FileSequence

pub fn new(files: &[(PathBuf, u64)], auto_set_len: bool) -> Self

pub fn set_file_size(&mut self, file_index: usize, file_size: u64)

Update a single file size.

pub fn get_current_file(&self) -> &Option<File>

Return current_file.

pub fn get_files(&self) -> &Vec<(PathBuf, u64)>

Return files.

pub fn len(&self) -> u64

Return the combined file size of all files.

pub fn is_empty(&self) -> bool

Return true if the FileSequence contains no file.

pub fn subset_len(&self, files: HashSet<PathBuf>) -> u64

Return the combined file size of all files.

pub fn get_file_position(&self, file_index: usize) -> u64

Compute the starting position of the file (in bytes) by suming up the size of the previous files.

async fn open_next_file(&mut self) -> Result<()>

Open the file at (current_file_index + 1). If no file is currently open (current_file_index is None), it opens the first file.

async fn open_file(&mut self, file_index: usize) -> Result<()>

Open the file at file_index.