Module darkfi::net::hosts

Hosts are a list of network addresses used when establishing an outbound connection. Hosts are shared across the network through the address protocol. When attempting to connect, a node will loop through addresses in the host store until it finds ones to connect to.

Structs

Hosts

Manages a store of network addresses.

Functions

filter_invalid 🔒

Auxiliary function to filter invalid(unresolvable) hosts.

filter_localnet 🔒

Auxiliary function to filter localnet hosts.

filter_non_resolving 🔒

Filters input_addrs keys to whatever has at least one IpAddr that is the same as connection_addr’s IP address. Skips .onion domains.

is_valid_onion 🔒

Validate a given .onion address. Currently it just checks that the length and encoding are ok, and does not do any deeper check. Should be fixed in the future.

Type Definitions

HostsPtr

Pointer to hosts class.