Module darkfi::net::hosts::store

The main interface for interacting with the hostlist. Contains the following:

Hosts: the main parent class that manages HostRegistry and HostContainer. It is also responsible for filtering addresses before writing to the hostlist.

HostRegistry: A locked HashMap that maps peer addresses onto mutually exclusive states (HostState). Prevents race conditions by dictating a strict flow of logically acceptable states.

HostContainer: A wrapper for the hostlists. Each hostlist is represented by a HostColor, which can be Grey, White, Gold or Black. Exposes a common interface for hostlist queries and utilities.

HostColor: White hosts have been seen recently. Gold hosts we have been able to establish a connection to. Grey hosts are recently received hosts that are periodically refreshed using the greylist refinery. Black hosts are considered hostile and are strictly avoided for the duration of the program. Dark hosts are hosts that do not match our transports, but that we continue to share with other peers. They are otherwise ignored.

HostState: a set of mutually exclusive states that can be Insert, Refine, Connect, Suspend or Connected. The state is None when the corresponding host has been removed from the HostRegistry.

Structs

HostContainer

A Container for managing Grey, White, Gold and Black hostlists. Exposes a common interface for writing to and querying hostlists.

Hosts

Main parent class for the management and manipulation of hostlists. Keeps track of hosts and their current state via the HostRegistry, and stores hostlists and associated methods in the HostContainer. Also operates two subscribers to notify other parts of the code base when new channels have been created or new hosts have been added to the hostlist.

Enums

HostColor

HostState

HostState is a set of mutually exclusive states that can be Insert, Refine, Move, Connect, Suspend or Connected. The state is None when the corresponding host has been removed from the HostRegistry.

Constants

DARKLIST_MAX_LEN 🔒

GREYLIST_MAX_LEN 🔒

LOCAL_HOST_STRS

WHITELIST_MAX_LEN 🔒

Type Aliases

HostRegistry

Keeps track of hosts and their current state. Prevents race conditions where multiple threads are simultaneously trying to change the state of a given host.

HostsPtr

Atomic pointer to hosts object