Module darkfi::net::message

Defines how to decode generic messages as well as implementing the common network messages that are sent between nodes as described by the Protocol submodule.

Implements a type called Packet which is the base message type. Packets are converted into messages and passed to an event loop.

Structs

AddrsMessage

Sends address information to inbound connection. Response to GetAddrs message.

ExtAddrsMessage

Sends external address information to inbound connection.

GetAddrsMessage

Requests address of outbound connection.

Packet

Packets are the base type read from the network. Converted to messages and passed to event loop.

PingMessage

Outbound keep-alive message.

PongMessage

Inbound keep-alive message.

VerackMessage

Sends version information to inbound connection. Response to VersionMessage.

VersionMessage

Requests version information of outbound connection.

Constants

MAGIC_BYTES 🔒

Traits

Message

Generic message template.

Functions

read_packet

Reads and decodes an inbound payload.

send_packet

Sends an outbound packet by writing data to TCP stream.