Module darkfi::net::message_subscriber

Generic publish/subscribe class that can dispatch any kind of message to a subscribed list of dispatchers. Dispatchers subscribe to a single message format of any type. This is a generalized version of the simple publish-subscribe class in system::Subscriber.

Message Subsystem also enables the creation of new message subsystems, adding new dispatchers and clearing inactive channels.

Message Subsystem maintains a list of dispatchers, which is a generalized version of a subscriber. Pub-sub is called on dispatchers through the functions ‘subscribe’ and ‘notify’. Whereas system::Subscriber only allows messages of a single type, dispatchers can handle any kind of message. This generic message is called a payload and is processed and decoded by the Message Dispatcher.

The Message Dispatcher is a class of subscribers that implements a generic trait called Message Dispatcher Interface, which allows us to process any kind of payload as a message.

Structs

MessageDispatcher 🔒

A dispatchers that is unique to every Message. Maintains a list of subscribers that are subscribed to that unique Message type and handles sending messages across these subscriptions.

MessageSubscription

Handles message subscriptions through a subscription ID and a receiver channel.

MessageSubsystem

Generic publish/subscribe class that maintains a list of dispatchers. Dispatchers transmit messages to subscribers and are specific to one message type.

Traits

MessageDispatcherInterface 🔒

Generic interface for message dispatcher.

Type Definitions

MessageResult 🔒

MessageSubscriptionId

64bit identifier for message subscription.