Module darkfi::net::p2p

P2P provides all core functionality to interact with the peer-to-peer network.

Used to create a network, to start and run it, to broadcast messages across all channels, and to manage the channel store.

The channel store is a hashmap of channel address that we can use to add and remove channels or check whether a channel is already is in the store.

Structs

P2p

Top level peer-to-peer networking interface.

Enums

P2pState 🔒

Type Definitions

ConnectedChannels

List of connected channels.

P2pPtr

Atomic pointer to p2p interface.

PendingChannels

List of channels that are awaiting connection.