Module darkfi::net::protocol::protocol_ping

Protocol for ping-pong keep-alive messages. Implements ping message and pong response. These messages are like the network heartbeat- they are sent continually between nodes, to ensure each node is still alive and active. Ping-pong messages ensure that the network doesn’t time out.

Protocol starts by creating a subscription to ping and pong messages. Then it starts a loop with a timer and runs ping-pong in the task manager. It sends out a ping and waits for pong reply. Then waits for ping and replies with a pong.

Structs

ProtocolPing

Defines ping and pong messages.