darkfi::net::session

Module direct_session

Direct connections session. Manages the creation of direct sessions. Used to create a direct session and to stop and start the session.

A direct session is a type of outbound session in which a protocol can open a temporary channel (stopped after used) to a peer. Direct sessions do not loop continually, once stopped the session will not try to reopen a connection.

ChannelBuilder is used to create new direct connections.

If there is no slots in the outbound session, the direct session can optionnally handle peer discovery.

Structs

ChannelBuilder

DirectSession

Defines direct connections session.

PeerDiscovery 🔒

PeerDiscovery process for that sends GetAddrs messages to a random whitelist or greylist host (creating a channel in the direct session). If it’s unsuccessful after two attempts, connect to our seed nodes and perform SeedSyncSession.

Type Aliases

DirectSessionPtr