Module darkfi::net::session::inbound_session

Inbound connections session. Manages the creation of inbound sessions. Used to create an inbound session and start and stop the session.

Class consists of 3 pointers: a weak pointer to the p2p parent class, an acceptor pointer, and a stoppable task pointer. Using a weak pointer to P2P allows us to avoid circular dependencies.

Structs

InboundInfo 🔒

InboundSession

Defines inbound connections session.