Module darkfi::net::session::outbound_session

Outbound connections session. Manages the creation of outbound sessions. Used to create an outbound session and stop and start the session.

Class consists of a weak pointer to the p2p interface and a vector of outbound connection slots. Using a weak pointer to p2p allows us to avoid circular dependencies. The vector of slots is wrapped in a mutex lock. This is switched on everytime we instantiate a connection slot and insures that no other part of the program uses the slots at the same time.

Structs

OutboundInfo 🔒

OutboundSession

Defines outbound connections session.

Enums

OutboundState 🔒