Module darkfi::net::session::seedsync_session

Seed sync session creates a connection to the seed nodes specified in settings. A new seed sync session is created every time we call p2p::start(). The seed sync session loops through all the configured seeds and tries to connect to them using a Connector. Seed sync either connects successfully, fails with an error or times out.

If a seed node connects successfully, it runs a version exchange protocol, stores the channel in the p2p list of channels, and disconnects, removing the channel from the channel list.

The channel is registered using Session trait method, register_channel(). This invokes the Protocol Registry method attach(). Usually this returns a list of protocols that we loop through and start. In this case, attach() uses the bitflag selector to identify seed sessions and exclude them.

The version exchange occurs inside register_channel(). We create a handshake task that runs the version exchange with the function perform_handshake_protocols(). This runs the version exchange protocol, stores the channel in the p2p list of channels, and subscribes to a stop signal.

Structs

SeedSyncSession

Defines seed connections session.