Module darkfi::net::transport

Network transport implementations.

Modules

tcp 🔒

tor 🔒

unix 🔒

upgrade_tls 🔒

Structs

TcpTransport

TlsUpgrade

TorTransport

Implements communication through the tor proxy service.

UnixTransport

Enums

TransportName

Traits

Transport

The Transport trait serves as a base for implementing transport protocols. Base transports can optionally be upgraded with TLS in order to support encryption. The implementation of our TLS authentication can be found in the upgrade_tls module.

TransportListener

Used as wrapper for listener used by Transport trait

TransportStream

Used as wrapper for stream used by Transport trait

Functions

socket_addr_to_url 🔒

A helper function to convert SocketAddr to Url and add scheme