Struct darkfi::raft::consensus::Raft

pub struct Raft<T> {

    id: NodeId,
    pub(super) role: Role,
    pub(super) current_leader: NodeId,
    pub(super) votes_received: Vec<NodeId>,
    pub(super) sent_length: MapLength,
    pub(super) acked_length: MapLength,
    pub(super) nodes: Arc<Mutex<HashMap<NodeId, i64>>>,
    pub(super) last_term: u64,
    pub(super) last_heartbeat: i64,
    p2p_sender: (Sender<NetMsg>, Receiver<NetMsg>),
    msgs_channel: (Sender<T>, Receiver<T>),
    commits_channel: (Sender<T>, Receiver<T>),
    datastore: DataStore<T>,
    seen_msgs: Arc<Mutex<HashMap<String, i64>>>,
    pub(super) settings: RaftSettings,
    pending_msgs: Vec<T>,

}

Fields
id: NodeIdrole: Rolecurrent_leader: NodeIdvotes_received: Vec<NodeId>sent_length: MapLengthacked_length: MapLengthnodes: Arc<Mutex<HashMap<NodeId, i64>>>last_term: u64last_heartbeat: i64p2p_sender: (Sender<NetMsg>, Receiver<NetMsg>)msgs_channel: (Sender<T>, Receiver<T>)commits_channel: (Sender<T>, Receiver<T>)datastore: DataStore<T>seen_msgs: Arc<Mutex<HashMap<String, i64>>>settings: RaftSettingspending_msgs: Vec<T>
Implementations

impl<T: Decodable + Encodable + Clone> Raft<T>

pub fn new(
settings: RaftSettings,
seen_msgs: Arc<Mutex<HashMap<String, i64>>>
) -> Result<Self>

pub async fn run(
&mut self,
p2p: P2pPtr,
p2p_recv_channel: Receiver<NetMsg>,
executor: Arc<Executor<'_>>,
stop_signal: Receiver<()>
) -> Result<()>

Run raft consensus and wait stop_signal channel to terminate

pub fn receiver(&self) -> Receiver<T>

Return async receiver channel which can be used to receive T Messages from raft consensus

pub fn sender(&self) -> Sender<T>

Return async sender channel which can be used to broadcast T Messages to raft consensus

pub fn id(&self) -> NodeId

Return the raft node id

async fn send_id_msg(&self) -> Result<()>

async fn broadcast_msg(&mut self, msg: &T, msg_id: Option<u64>) -> Result<()>

async fn handle_method(&mut self, msg: NetMsg) -> Result<()>

pub(super) async fn send(
&self,
recipient_id: Option<NodeId>,
payload: &[u8],
method: NetMsgMethod,
msg_id: Option<u64>
) -> Result<()>

pub(super) fn reset_last_term(&mut self) -> Result<()>

pub(super) fn set_current_term(&mut self, i: &u64) -> Result<()>

pub(super) fn set_voted_for(&mut self, i: &Option<NodeId>) -> Result<()>

pub(super) async fn push_commit(&mut self, commit: &[u8]) -> Result<()>

pub(super) fn push_log(&mut self, log: &Log) -> Result<()>

pub(super) fn push_logs(&mut self, logs: &Logs) -> Result<()>

pub(super) fn current_term(&self) -> Result<u64>

pub(super) fn voted_for(&self) -> Result<Option<NodeId>>

pub(super) fn commits_len(&self) -> u64

fn logs(&self) -> Result<Logs>

pub(super) fn logs_len(&self) -> u64

fn last_log(&self) -> Result<Option<Log>>

pub(super) fn get_log(&self, index: u64) -> Result<Log>

pub(super) fn slice_logs_from(&self, index: u64) -> Result<Option<Logs>>

pub(super) fn slice_logs_to(&self, index: u64) -> Result<Logs>

impl<T: Decodable + Encodable + Clone> Raft<T>

pub(super) async fn send_vote_request(&mut self) -> Result<()>

pub(super) async fn receive_vote_response(
&mut self,
vr: VoteResponse
) -> Result<()>

impl<T: Decodable + Encodable + Clone> Raft<T>

pub(super) async fn receive_vote_request(
&mut self,
vr: VoteRequest
) -> Result<()>

pub(super) async fn receive_log_request(&mut self, lr: LogRequest) -> Result<()>

async fn append_log(
&mut self,
prefix_len: u64,
leader_commit: u64,
suffix: &Logs
) -> Result<()>

impl<T: Decodable + Encodable + Clone> Raft<T>

pub(super) async fn send_heartbeat(&mut self) -> Result<()>

async fn update_logs(&mut self, node_id: &NodeId) -> Result<()>

pub(super) async fn receive_log_response(
&mut self,
lr: LogResponse
) -> Result<()>

fn acks(&self, nodes: HashMap<NodeId, i64>, length: u64) -> HashMap<NodeId, i64>

async fn commit_log(&mut self) -> Result<()>