darkfi::rpc
Module server

Server-side JSON-RPC implementation

Traits

RequestHandler

Asynchronous trait implementing a handler for incoming JSON-RPC requests.

Functions

accept

Accept function that should run inside a loop for accepting incoming JSON-RPC requests and passing them to the RequestHandler.

handle_request 🔒

Auxiliary function to handle a request in the background.

listen_and_serve

Start a JSON-RPC server bound to the givven accept URL and use the given RequestHandler to handle incoming requests.

run_accept_loop 🔒

Wrapper function around accept() to take the incoming connection and pass it forward.