darkfi::runtime::import::util
Function set_return_data

pub(crate) fn set_return_data(
    ctx: FunctionEnvMut<'_, Env>,
    ptr: WasmPtr<u8>,
    len: u32,
) -> i64

Writes data to the contract_return_data field of Env. The data will be read from ptr at a memory offset specified by len.

Returns SUCCESS on success, otherwise returns an error code corresponding to a [ContractError].

Permissions: metadata, exec