darkfi::runtime

Module vm_runtime

Main WASM VM runtime implementation

Structs

Env

The WASM VM runtime environment instantiated for every smart contract that runs.

Runtime

Define a wasm runtime.

Enums

ContractSection

Constants

GAS_LIMIT

Gas limit for a single contract call (Single WASM instance)

MEMORY 🔒

Name of the wasm linear memory in our guest module