darkfi::system

Module timeout

Async timeout implementations

Structs

Timeout

TimeoutError

An error returned when a future times out.

TimeoutFuture

A future that times out after a duration of time.

Functions

io_timeout

Awaits an I/O future or times out after a duration of time.

timeout

Awaits a future or times out after a duration of time.