darkfi::util

Module logger

Logging utilities

Macros

verbose

Structs

ChannelWriter

Wrapper structure over a smol channel sender to use as a logs writer for the interactive shell.

EventFormatter

Formats tracing events for output, with support for custom log levels.

EventTimeFormatter

Formats event timestamps as HH:MM:SS for tracing output.

LevelFormatter 🔒

A formatter for the custom Level

TargetFilter

A tracing-subscriber layer that filters events based on their target.

Enums

Level

A custom log level type.

Statics

TERMINAL_LAYER_HANDLE 🔒

Global singleton holding the tracing terminal layer handle.

Traits

TerminalWriter

Trait definitions for cleaner code.

Functions

file_field_formatter

Formats event fields for file logging, hiding level fields like verbose.

set_terminal_writer

Auxiliary function to set the tracing terminal layer writer.

setup_logging

Helper for setting up logging for bins.

setup_test_logger

Helper for setting up terminal logging for tests.

terminal_field_formatter

Formats event fields for terminal output, hiding level fields like verbose.