darkfi::util::logger

Struct TargetFilter

pub struct TargetFilter {
    ignored_targets: Vec<String>,
    allowed_targets: Vec<String>,
    target_levels: Vec<(String, Level)>,
    default_level: Level,
}

A tracing-subscriber layer that filters events based on their target.

Fields
ignored_targets: Vec<String>

Targets where logs should always be ignored.

allowed_targets: Vec<String>

If non-empty, only these targets are allowed.

target_levels: Vec<(String, Level)>

Per-target minimum level that must be met for an event to be logged.

default_level: Level

Default minimum level if no override is found

Implementations

impl TargetFilter

pub fn parse_targets(self, log_targets: String) -> Self

Create a TargetFilter by parsing allowed_targets and ignored_targets from a string in the format target1,target2 or !target1,!target2

pub fn with_verbosity(self, verbosity_level: u8) -> Self

Overrides default level with verbosity level.

pub fn allow_targets<I, S>(self, targets: I) -> Self
where I: IntoIterator<Item = S>, S: AsRef<str>,

Adds multiple allowed targets.

pub fn ignore_targets<I, S>(self, targets: I) -> Self
where I: IntoIterator<Item = S>, S: AsRef<str>,

Adds multiple ignored targets.

pub fn targets_level<I, S>(self, targets: I, level: Level) -> Self
where I: IntoIterator<Item = S>, S: AsRef<str>,

Assign a log level filter for multiple targets.

pub fn default_level(self, level: Level) -> Self

Override default level filter.

fn filter<S: Subscriber>( &self, metadata: &Metadata<'_>, _ctx: &Context<'_, S>, ) -> bool

Helper method to implement the filtering logic.