darkfi::util
Module time

Time utilities

Structs

Date

DateTime

Represents a UTC DateTime with individual fields for date and time components.

NanoTimestamp

Timestamp

Wrapper struct to represent system timestamps.

Enums

DateFormat

Constants

DAYS_IN_MONTHS 🔒

Represents the number of days in each month for both leap and non-leap years.

MIN_IN_HOUR 🔒

SECS_IN_DAY 🔒

SECS_IN_HOUR 🔒

Functions

fmt_duration

Formats a Duration into a user-friendly format using days, hours, minutes, and seconds, and returns the formatted string.

timestamp_to_date