darkfi::validator
Module consensus

DarkFi consensus module

Structs

Consensus

This struct represents the information required by the consensus algorithm

Fork

Struct representing a forked blockchain state.

Proposal

This struct represents a block proposal, used for consensus.

Constants

GAS_LIMIT_MULTIPLIER_UNPROPOSED_TXS 🔒

Multiplier used to calculate the gas limit for unproposed transactions

GAS_LIMIT_UNPROPOSED_TXS

Gas limit for unproposed transactions

GAS_TX_AVG 🔒

Average amount of gas consumed during transaction execution, derived by the Gas Analyzer