Module darkfi::validator::pid

DarkFi consensus PID controller

Structs

EPSILON ğŸ”’

FIELD_P ğŸ”’

K1 ğŸ”’

K2 ğŸ”’

K3 ğŸ”’

KD ğŸ”’

KI ğŸ”’

KP ğŸ”’

MAX_F ğŸ”’

MIN_F ğŸ”’

Constants

P ğŸ”’

PID controller configuration

Functions

calculate_f ğŸ”’

Calculate the inverse probability f of becoming a block producer (winning the lottery) having all the tokens, and the feedback error, represented as Float10.

calculate_sigmas ğŸ”’

Return 2-term target approximation sigma coefficients, corresponding to provided f and total_tokens values.

slot_pid_output

Return 2-term target approximation sigma coefficients, alogn with the inverse probability f of becoming a block producer and the feedback error, corresponding to provided slot consensus state,