Module darkfi::validator::verification

Verification functions

Functions

validate_block

A block is considered valid when the following rules apply: 1. Block version is correct for its height 2. Parent hash is equal to the hash of the previous block 3. Block height increments previous block height by 1 4. Timestamp is valid based on PoWModule validation 5. Block hash is valid based on PoWModule validation Additional validity rules can be applied.

validate_blockchain

A blockchain is considered valid, when every block is valid, based on validate_block checks. Be careful as this will try to load everything in memory.

verify_block

Verify given BlockInfo, and apply it to the provided overlay

verify_genesis_block

Verify given genesis BlockInfo, and apply it to the provided overlay

verify_producer_signature

Verify block proposer signature, using the proposal transaction signature as signing key over blocks header hash.

verify_producer_transaction

Verify WASM execution, signatures, and ZK proofs for a given producer Transaction, and apply it to the provided overlay. Returns transaction signature public key. Additionally, append its hash to the provided Merkle tree.

verify_proposal

Verify given Proposal against provided consensus state, A proposal is considered valid when the following rules apply: 1. Proposal hash matches the actual block one 2. Block transactions don’t exceed set limit 3. Block is valid Additional validity rules can be applied.

verify_transaction

Verify WASM execution, signatures, and ZK proofs for a given Transaction, and apply it to the provided overlay. Additionally, append its hash to the provided Merkle tree.

verify_transactions

Verify a set of Transaction in sequence and apply them if all are valid. In case any of the transactions fail, they will be returned to the caller as an error. If all transactions are valid, the function will return the accumulated gas used from all the transactions. Additionally, their hash is appended to the provided Merkle tree.