darkfi::validator::verification

Function validate_block

pub fn validate_block(
    block: &BlockInfo,
    previous: &BlockInfo,
    module: &PoWModule,
) -> Result<()>

Validate provided block according to set rules.

A block is considered valid when the following rules apply: 1. Block version is correct for its height 2. Previous hash is equal to the hash of the provided previous block 3. Block height increments previous block height by 1 4. Timestamp is valid based on PoWModule validation 5. Block header Proof of Work data are valid 6. Block hash is valid based on PoWModule validation Additional validity rules can be applied.