darkfi::validator::verification
Function validate_blockchain

pub fn validate_blockchain(
    blockchain: &Blockchain,
    pow_target: u32,
    pow_fixed_difficulty: Option<BigUint>,
) -> Result<()>

A blockchain is considered valid, when every block is valid, based on validate_block checks. Be careful as this will try to load everything in memory.