darkfi::validator::verification
Function verify_fork_proposal

pub async fn verify_fork_proposal(
    fork: &Fork,
    proposal: &Proposal,
    verify_fees: bool,
) -> Result<()>

Verify given Proposal against provided fork state.

A proposal is considered valid when the following rules apply: 1. Proposal hash matches the actual block one 2. Block is valid Additional validity rules can be applied.