Module darkfi::zk::halo2

Modules

plonk

This module provides an implementation of a variant of (Turbo)PLONK that is designed specifically for the polynomial commitment scheme described in the Halo paper.

Structs

Advice

An advice column

AssignedCell

An assigned cell.

Column

A column with an index and type

Value

A value that might exist within a circuit.

Enums

Assigned

A value assigned to a cell within a circuit.

Traits

Field

This trait represents an element of a field.

Layouter

A layout strategy within a circuit. The layouter is chip-agnostic and applies its strategy to the context and config it is given.