Crate darkfi

Re-exports

pub use error::ClientFailed;

pub use error::ClientResult;

pub use error::Error;

pub use error::Result;

Modules

blockchain

dht

error

event_graph

geode

Chunk-based file storage implementation. This is a building block for a DHT or something similar.

net

rpc

runtime

system

tx

util

validator

zk

zkas

src/zkas is the library holding the zkas toolchain, consisting of a lexer, parser, static/semantic analyzers, a binary compiler, and a binary decoder.

Macros

async_daemonize

This macro is used for a standard way of daemonizing darkfi binaries with TOML config file configuration, and argument parsing.

cli_desc

impl_dht_node_defaults

Implements default Hash, PartialEq, and Eq for a struct implementing DhtNode

impl_p2p_message

verbose

Constants

ANSI_LOGO