Crate darkfi

Re-exports

pub use error::ClientFailed;

pub use error::ClientResult;

pub use error::Error;

pub use error::Result;

pub use error::VerifyFailed;

pub use error::VerifyResult;

Modules

blockchain

consensus

dht

error

event_graph

net

raft

rpc

runtime

system

tx

util

wallet

zk

zkas

src/zkas is the library holding the zkas toolchain, consisting of a lexer, parser, static/semantic analyzers, a binary compiler, and a binary decoder.

Macros

async_daemonize

This macro is used for a standard way of daemonizing darkfi binaries with TOML config file configuration, and argument parsing. It also spawns a multithreaded async executor and passes it into the given function.

cli_desc

Constants

ANSI_LOGO