Module darkfi::blockchain

Re-exports

pub use block_store::Block;

pub use block_store::BlockDifficulty;

pub use block_store::BlockInfo;

pub use block_store::BlockStore;

pub use block_store::BlockStoreOverlay;

pub use block_store::SLED_BLOCK_DIFFICULTY_TREE;

pub use block_store::SLED_BLOCK_ORDER_TREE;

pub use block_store::SLED_BLOCK_STATE_DIFF_TREE;

pub use block_store::SLED_BLOCK_TREE;

pub use header_store::Header;

pub use header_store::HeaderHash;

pub use header_store::HeaderStore;

pub use header_store::HeaderStoreOverlay;

pub use header_store::SLED_HEADER_TREE;

pub use header_store::SLED_SYNC_HEADER_TREE;

pub use tx_store::TxStore;

pub use tx_store::TxStoreOverlay;

pub use tx_store::SLED_PENDING_TX_ORDER_TREE;

pub use tx_store::SLED_PENDING_TX_TREE;

pub use tx_store::SLED_TX_LOCATION_TREE;

pub use tx_store::SLED_TX_TREE;

pub use contract_store::ContractStore;

pub use contract_store::ContractStoreOverlay;

pub use contract_store::SLED_BINCODE_TREE;

pub use contract_store::SLED_CONTRACTS_TREE;

Modules

block_store

Block related definitions and storage implementations

contract_store

Contracts and Wasm storage implementations

header_store

Header definition and storage implementation

tx_store

Transactions related storage implementations

Structs

Blockchain

Structure holding all sled trees that define the concept of Blockchain.

BlockchainOverlay

Overlay structure over a Blockchain instance.

Functions

parse_record

Parse a sled record in the form of a tuple (key, value).

parse_u32_key_record

Parse a sled record with a u32 key in the form of a tuple (key, value).

parse_u64_key_record

Parse a sled record with a u64 key in the form of a tuple (key, value).

Type Aliases

BlockchainOverlayPtr

Atomic pointer to blockchain overlay.

SledDbOverlayPtr

Atomic pointer to sled db overlay.