Module darkfi::blockchain::block_store

Block related definitions and storage implementations

Structs

Block

This struct represents a tuple of the form (header, txs, signature). The header and transactions are stored as hashes, serving as pointers to the actual data in the sled database. NOTE: This struct fields are considered final, as it represents a blockchain block.

BlockDifficulty

Auxiliary structure used to keep track of block PoW difficulty information. Note: we only need height cummulative difficulty, but we also keep its actual difficulty, so we can verify the sequence and/or know specific block height difficulty, if ever needed.

BlockDifficultyStore

The BlockDifficultyStore is a sled tree storing the difficulty information of the blockchain’s blocks, where the key is the block height number, and the value is the blocks’ hash. BlockDifficultyStore can be queried with this height number.

BlockDifficultyStoreOverlay

Overlay structure over a BlockDifficultyStore instance.

BlockInfo

Structure representing full block data, acting as a wrapper struct over Block, enabling us to include more information that might be used in different block version, without affecting the original struct.

BlockOrder

Auxiliary structure used to keep track of blocks order.

BlockOrderStore

The BlockOrderStore is a sled tree storing the order of the blockchain’s blocks, where the key is the order number, and the value is the blocks’ hash. BlockOrderStore can be queried with this order number.

BlockOrderStoreOverlay

Overlay structure over a BlockOrderStore instance.

BlockRanks

Auxiliary structure used to keep track of block ranking information. Note: we only need height cummulative ranks, but we also keep its actual ranks, so we can verify the sequence and/or know specific block height ranks, if ever needed.

BlockStore

The BlockStore is a sled tree storing all the blockchain’s blocks where the key is the blocks’ hash, and value is the serialized block.

BlockStoreOverlay

Overlay structure over a BlockStore instance.

Constants

SLED_BLOCK_DIFFICULTY_TREE 🔒

BlockDifficulty sled tree

SLED_BLOCK_ORDER_TREE 🔒

BlockOrder sled tree

SLED_BLOCK_TREE 🔒

Block sled tree

Functions

append_tx_to_merkle_tree

Auxiliary function to append a transaction to a Merkle tree.