Module darkfi::blockchain::block_store

Block related definitions and storage implementations

Structs

Block

This struct represents a tuple of the form (header, txs, signature).

BlockDifficulty

Auxiliary structure used to keep track of block PoW difficulty information.

BlockInfo

Structure representing full block data.

BlockOrder

Auxiliary structure used to keep track of blocks order.

BlockRanks

Auxiliary structure used to keep track of block ranking information.

BlockStore

The BlockStore is a structure representing all sled trees related to storing the blockchain’s blocks information.

BlockStoreOverlay

Overlay structure over a BlockStore instance.

Constants

SLED_BLOCK_DIFFICULTY_TREE

SLED_BLOCK_ORDER_TREE

SLED_BLOCK_STATE_DIFF_TREE

SLED_BLOCK_TREE

Functions

append_tx_to_merkle_tree

Auxiliary function to append a transaction to a Merkle tree.