Module darkfi::blockchain::block_store

Structs

BlockOrderStore

The BlockOrderStore is a sled tree storing the order of the blockchain’s slots, where the key is the slot uid, and the value is the blocks’ hash. BlockStore can be queried with this hash.

BlockStore

The BlockStore is a sled tree storing all the blockchain’s blocks where the key is the blocks’ hash, and value is the serialized block.

HeaderStore

The HeaderStore is a sled tree storing all the blockchain’s blocks’ headers where the key is the headers’ hash, and value is the serialized header.

Constants

SLED_BLOCK_ORDER_TREE 🔒

SLED_BLOCK_TREE 🔒

SLED_HEADER_TREE 🔒