Module darkfi::consensus::constants

Constants

Structs

FLOAT10_EPSILON

FLOAT10_FIVE

FLOAT10_NEG_ONE

FLOAT10_NEG_TWO

FLOAT10_NINE

FLOAT10_ONE

FLOAT10_TEN

FLOAT10_THREE

FLOAT10_TWO

FLOAT10_ZERO

KD

KI

KP

MAINNET_BOOTSTRAP_TIMESTAMP

Bootstrap timestamp for the mainnet chain

MAINNET_GENESIS_HASH_BYTES

Genesis hash for the mainnet chain

MAINNET_GENESIS_TIMESTAMP

Genesis timestamp for the mainnet chain

MAINNET_INITIAL_DISTRIBUTION

Total sum of initial staking coins for the mainnet chain

MAX_DER

MAX_F

MIN_DER

MIN_F

PID_OUT_STEP

TESTNET_BOOTSTRAP_TIMESTAMP

Bootstrap timestamp for the testnet chain

TESTNET_GENESIS_HASH_BYTES

Genesis hash for the testnet chain

TESTNET_GENESIS_TIMESTAMP

Genesis timestamp for the testnet chain

TESTNET_INITIAL_DISTRIBUTION

Total sum of initial staking coins for the testnet chain

Constants

BLOCK_INFO_MAGIC_BYTES

Block info magic bytes

BLOCK_MAGIC_BYTES

Block magic bytes

BLOCK_VERSION

Block version number

CONSENSUS_COIN_COL

CONSENSUS_COIN_TABLE

EPOCH_LENGTH

Number of slots in one epoch

FINAL_SYNC_DUR

Finalization sync period duration (should be >=2/3 of slot time)

F_HISTORY_LOG

GENESIS_TOTAL_STAKE

LEADER_HISTORY_LOG

LEADER_PROOF_K

Leader proofs k for zk proof rows (rows=2^k)

LOTTERY_HEAD_START

LOTTERY_HISTORY_LOG

P

PI_COMMITMENT_ROOT

PI_COMMITMENT_X_INDEX

PI_COMMITMENT_Y_INDEX

PI_MU_RHO_INDEX

PI_MU_Y_INDEX

PI_NULLIFIER_INDEX

PI_SIGMA1_INDEX

PI_SIGMA2_INDEX

PRF_NULLIFIER_PREFIX

RADIX_BITS

REWARD

Block leader reward

SLOT_TIME

Slot time in seconds

SYNC_MAX_RETRIES

Max resync retries

SYNC_RETRIES_DURATION

Max resync retries duration in epochs

TXS_CAP

Transactions included in a block cap

Statics

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

Functions

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’