Module darkfi::event_graph::proto

P2P protocol implementation for the Event Graph

Structs

EventPut

A P2P message representing publishing an event on the network

EventRep

A P2P message representing an event reply

EventReq

A P2P message representing an event request

ProtocolEventGraph

P2P protocol implementation for the Event Graph.

TipRep

A P2P message representing a reply for the peer’s DAG tips

TipReq

A P2P message representing a request for a peer’s DAG tips

Constants

MALICIOUS_THRESHOLD 🔒

Malicious behaviour threshold. If the threshold is reached, we will drop the peer from our P2P connection.