darkfi::event_graph

Module util

Utility functions

Constants

DAY 🔒

MilliSeconds in a day

Functions

days_since 🔒

Calculate the number of days since a given midnight timestamp.

generate_genesis

Generate a deterministic genesis event corresponding to the DAG’s configuration.

midnight_timestamp 🔒

Calculate the midnight timestamp given a number of days. If days is 0, calculate the midnight timestamp of today.

millis_until_next_rotation

Calculate the time in milliseconds until the next_rotation, given as a timestamp. next_rotation here represents a timestamp in UNIX epoch format.

next_rotation_timestamp

Calculate the timestamp of the next DAG rotation.

recreate_from_replayer_log

replayer_log 🔒