darkfi::net

Module message

Defines how to decode generic messages as well as implementing the common network messages that are sent between nodes as described by the protocol submodule.

Implements a type called Packet which is the base message type. Packets are converted into messages and passed to an event loop.

Structs

AddrsMessage

Sends address information to inbound connection.

GetAddrsMessage

Requests address of outbound connection.

PingMessage

Outbound keepalive message.

PongMessage

Inbound keepalive message.

SerializedMessage

Generic serialized message template.

VerackMessage

Sends version information to inbound connection. Response to VersionMessage.

VersionMessage

Requests version information of outbound connection.

Constants

ADDRS_MAX_BYTES

Addrs message fields size:

ADDRS_METERING_CONFIGURATION

GET_ADDRS_MAX_BYTES

GetAddrs message fields size:

GET_ADDRS_METERING_CONFIGURATION

MAX_COMMAND_LENGTH

Maximum command (message name) length in bytes.

PING_PONG_MAX_BYTES

Ping-Pong messages fields size:

PING_PONG_METERING_CONFIGURATION

For each message configs a threshold was calculated by taking the maximum number of messages in a 10 seconds window and multiply it by 2 not to be strict.

VERACK_MAX_BYTES

Verack message fields size:

VERACK_METERING_CONFIGURATION

VERSION_MAX_BYTES

Version message fields size:

VERSION_METERING_CONFIGURATION

Traits

Message

Generic message template.