Module darkfi::net::message

Defines how to decode generic messages as well as implementing the common network messages that are sent between nodes as described by the protocol submodule.

Implements a type called Packet which is the base message type. Packets are converted into messages and passed to an event loop.

Structs

AddrsMessage

Sends address information to inbound connection.

GetAddrsMessage

Requests address of outbound connecction.

PingMessage

Outbound keepalive message.

PongMessage

Inbound keepalive message.

SerializedMessage

Generic serialized message template.

VerackMessage

Sends version information to inbound connection. Response to VersionMessage.

VersionMessage

Requests version information of outbound connection.

Constants

MAGIC_BYTES 🔒

Traits

Message

Generic message template.