darkfi::net
Module p2p

P2P provides all core functionality to interact with the P2P network.

Used to create a network, to start and run it, to broadcast messages across all channels, and to manage the channel store.

The channel store is a hashmap of channel addresses that we can use to add and remove channels or check whether a channel is already in the store.

Structs

P2p

Toplevel peer-to-peer networking interface

Functions

broadcast_serialized_to 🔒

Auxiliary function to broadcast a serialized message concurrently to all given peers.

Type Aliases

P2pPtr

Atomic pointer to the p2p interface