darkfi::net::protocol
Module protocol_generic

Generic protocol to receive specified structure messages.

Acts as a simple message queue, where we listen for the specified structure message, and when one is received, we send it to the provided smol channel. Afterwards, we wait for an action signal, specifying whether or not we should propagate the message to rest nodes or skip it.

Structs

ProtocolGeneric

Defines generic messages protocol.

ProtocolGenericHandler

Defines a handler for generic protocol messages, consisting of a message receiver, action signal senders mapped by each channel ID, and a stoppable task to run the handler in the background.

Enums

ProtocolGenericAction

Defines generic messages protocol action signal.

Type Aliases

ProtocolGenericHandlerPtr