Module darkfi::net::session

Defines the interaction between nodes during a connection. Consists of an inbound session, which describes how to set up an incoming connection, and an outbound session, which describes setting up an outbound connection. Also describes the seed session, which is the type of connection used when a node connects to the network for the first time. Implements the session trait which describes the common functions across all sessions.

Re-exports

pub use inbound_session::InboundSession;

pub use manual_session::ManualSession;

pub use outbound_session::OutboundSession;

pub use seedsync_session::SeedSyncSession;

Modules

inbound_session

Inbound connections session. Manages the creation of inbound sessions. Used to create an inbound session and start and stop the session.

manual_session

outbound_session

Outbound connections session. Manages the creation of outbound sessions. Used to create an outbound session and stop and start the session.

seedsync_session

Seed sync session creates a connection to the seed nodes specified in settings. A new seed sync session is created every time we call p2p::start(). The seed sync session loops through all the configured seeds and tries to connect to them using a Connector. Seed sync either connects successfully, fails with an error or times out.

Constants

SESSION_ALL

SESSION_INBOUND

SESSION_MANUAL

SESSION_OUTBOUND

SESSION_SEED

Traits

Session

Session trait. Defines methods that are used across sessions. Implements registering the channel and initializing the channel by performing a network handshake.

Functions

remove_sub_on_stop 🔒

Removes channel from the list of connected channels when a stop signal is received.

Type Definitions

SessionBitflag

SessionWeakPtr