Module darkfi::net::session::inbound_session

Inbound connections session. Manages the creation of inbound sessions. Used to create an inbound session and start and stop the session.

Class consists of 3 pointers: a weak pointer to the p2p parent class, an acceptor pointer, and a stoppable task pointer. Using a weak pointer to P2P allows us to avoid circular dependencies.

Structs

InboundSession

Defines inbound connections session

PingSelfProcess

Periodically try to do a version exchange with our own external addresses. If the version exchange is successful, take a timestamp and save it along with the external addresses. Each address along with its timestamp (the last_seen data field) is sent in to other nodes in ProtocolAddr and ProtocolSeed.

Type Aliases

InboundSessionPtr