darkfi::net::session
Module outbound_session

Outbound connections session. Manages the creation of outbound sessions. Used to create an outbound session and to stop and start the session.

Class consists of a weak pointer to the p2p interface and a vector of outbound connection slots. Using a weak pointer to p2p allows us to avoid circular dependencies. The vector of slots is wrapped in a mutex lock. This is switched on every time we instantiate a connection slot and insures that no other part of the program uses the slots at the same time.

Structs

OutboundSession

Defines outbound connections session.

PeerDiscovery 🔒

Main PeerDiscovery process that loops through connected peers and sends out a GetAddrs when it is active. If there are no connected peers after two attempts, connect to our seed nodes and perform SeedSyncSession.

Slot 🔒

Traits

PeerDiscoveryBase

Defines a common interface for multiple peer discovery processes.

Type Aliases

OutboundSessionPtr