darkfi::net::session
Module refine_session

RefineSession manages the GreylistRefinery, which randomly selects entries on the greylist and updates them to whitelist if active,

GreylistRefinery makes use of a RefineSession method called handshake_node(), which uses a Connector to establish a Channel with a provided address, and then does a version exchange across the channel (perform_handshake_protocols). handshake_node() can either succeed, fail, or timeout.

Structs

GreylistRefinery

Periodically probes entries in the greylist.

RefineSession

Type Aliases

RefineSessionPtr