darkfi::net
Module transport

Network transports, holds implementations of pluggable transports. Exposes agnostic dialers and agnostic listeners.

Modules

nym ğŸ”’

Nym transport

socks5

SOCKS5 proxy client

tcp ğŸ”’

TCP transport

tls ğŸ”’

TLS upgrade mechanism

tor ğŸ”’

Tor transport

unix ğŸ”’

Unix socket transport

Macros

enforce_abspath ğŸ”’

enforce_hostport ğŸ”’

Structs

Dialer

A dialer that is able to transparently operate over arbitrary transports.

Listener

A listener that is able to transparently listen over arbitrary transports.

Enums

DialerVariant

Dialer variants

ListenerVariant

Listener variants

Traits

PtListener

Wrapper trait for async listeners

PtStream

Wrapper trait for async streams