Module darkfi::rpc::clock_sync

Clock sync utility module Clock sync module

Constants

EPOCH 🔒

NTP_ADDRESS 🔒

TODO: Loop through set of ntps, get their average response concurrenyly.

RETRIES 🔒

Clock sync parameters

Functions

check_clock

This is a very simple check to verify that the system time is correct. Retry loop is used in case discrepancies are found. If all retries fail, system clock is considered invalid. TODO: 1. Add proxy functionality in order not to leak connections

clock_check 🔒

ntp_request

Raw NTP request execution