darkfi::rpc
Module common

Common internal functions

Constants

INIT_BUF_SIZE 🔒

MAX_BUF_SIZE 🔒

READ_TIMEOUT 🔒

Functions

http_read_from_stream_request 🔒

Internal read function that reads from the active stream into a buffer. Performs HTTP POST request parsing. Returns the request body length.

http_read_from_stream_response 🔒

Internal read function that reads from the active stream into a buffer. Performs HTTP POST response parsing. Returns the response body length.

http_write_to_stream 🔒

Internal write function that writes a JSON-RPC object to the active stream. Sent as an HTTP response.

read_from_stream 🔒

Internal read function that reads from the active stream into a buffer. Reading stops upon reaching CRLF or LF, or when MAX_BUF_SIZE is reached.

write_to_stream 🔒

Internal write function that writes a JSON-RPC object to the active stream.