darkfi::runtime::import

Module db

Host functions for interacting with db backend

Structs

DbHandle

Internal wasm runtime API for sled trees

Functions

db_contains_key 🔒

Check if a database contains a given key.

db_del 🔒

Remove a key from the database.

db_get 🔒

Reads a value by key from the key-value store.

db_init 🔒

Create a new database instance for the calling contract.

db_lookup 🔒

Lookup a database handle from its name. If it exists, push it to the Vector of db_handles.

db_set 🔒

Set a value within the transaction.

zkas_db_set 🔒

Given a zkas circuit, create a VerifyingKey and insert them both into the db.