Module darkfi::runtime::import::util

Host functions for utilities

Functions

drk_log 🔒

Host function for logging strings.

get_blockchain_time 🔒

Will return current blockchain timestamp, defined as the last block’s timestamp.

get_last_block_height 🔒

Grabs last block from the Blockchain overlay and then copies its height to the VM’s object store.

get_object_bytes 🔒

Retrieve an object from the object store specified by the index idx. The object’s data is written to ptr.

get_object_size 🔒

Returns the size (number of bytes) of an object in the object store specified by index idx.

get_verifying_block_height 🔒

Will return current runtime configured verifying block height number

get_verifying_block_height_epoch 🔒

Will return current runtime configured verifying block height epoch number

put_object_bytes 🔒

Appends a new object to the Env objects store. The data for the object is read from ptr.

set_return_data 🔒

Writes data to the contract_return_data field of Env. The data will be read from ptr at a memory offset specified by len.