Module darkfi::system

Re-exports

pub use condvar::CondVar;

pub use stoppable_task::StoppableTask;

pub use stoppable_task::StoppableTaskPtr;

pub use publisher::Publisher;

pub use publisher::PublisherPtr;

pub use publisher::Subscription;

pub use timeout::io_timeout;

Modules

condvar

Condition variable which allows a task to block until woken up

publisher

Simple broadcast (publish-subscribe) class

stoppable_task

Implementation of async background task spawning which are stoppable using channel signalling.

timeout

Async timeout implementations

Functions

msleep

Sleep for any number of milliseconds.

run_until_completion

Run a task until it has fully completed, irrespective of whether the parent task still exists.

sleep

Sleep for any number of seconds.

sleep_forever

Type Aliases

ExecutorPtr