darkfi

Module tx

Macros

zip 🔒

Structs

ContractCallLeaf

Auxiliarry structure containing all the information required to execute a contract call.

Transaction

A Transaction contains an arbitrary number of ContractCall objects, along with corresponding ZK proofs and Schnorr signatures.

TransactionBuilder

Auxiliary structure to build a full Transaction using DarkTree to order everything.

Constants

MAX_TX_CALLS

MIN_TX_CALLS

Calls tree bounds definitions