Module darkfi::validator::verification

Verification functions

Functions

apply_producer_transaction 🔒

Apply given producer Transaction to the provided overlay, without formal verification. Returns transaction signature public key. Additionally, append its hash to the provided Merkle tree.

apply_transaction 🔒

Apply given Transaction to the provided overlay. Additionally, append its hash to the provided Merkle tree.

apply_transactions 🔒

Apply given set of Transaction in sequence, without formal verification. In case any of the transactions fail, they will be returned to the caller as an error. Additionally, their hash is appended to the provided Merkle tree.

validate_block

Validate provided block according to set rules.

validate_blockchain

A blockchain is considered valid, when every block is valid, based on validate_block checks. Be careful as this will try to load everything in memory.

verify_block

Verify given BlockInfo, and apply it to the provided overlay.

verify_checkpoint_block

Verify given checkpoint BlockInfo, and apply it to the provided overlay.

verify_genesis_block

Verify given genesis BlockInfo, and apply it to the provided overlay.

verify_producer_signature

Verify block proposer signature, using the producer transaction signature as signing key over blocks header hash.

verify_producer_transaction

Verify provided producer Transaction.

verify_proposal

Verify given Proposal against provided consensus state.

verify_transaction

Verify WASM execution, signatures, and ZK proofs for a given Transaction, and apply it to the provided overlay. Additionally, append its hash to the provided Merkle tree.

verify_transactions

Verify a set of Transaction in sequence and apply them if all are valid.