Module darkfi::validator::verification

Verification functions

Functions

verify_block

Verify given BlockInfo, and apply it to the provided overlay

verify_genesis_block

Verify given genesis BlockInfo, and apply it to the provided overlay

verify_pos_proposal

Verify given PoS Proposal against provided consensus state, A proposal is considered valid when the following rules apply: 1. Consensus(node) has not started current slot finalization 2. Proposal refers to current slot 3. Proposal hash matches the actual block one 4. Block transactions don’t exceed set limit 5. If proposal extends a known fork, verify block slots correspond to the fork hot/live ones 6. Block is valid Additional validity rules can be applied.

verify_pow_proposal

Verify given PoW Proposal against provided consensus state, A proposal is considered valid when the following rules apply: 1. Proposal hash matches the actual block one 2. Block transactions don’t exceed set limit 3. If proposal extends a known fork, verify block’s slot correspond to the fork hot/live/next one 4. Block is valid Additional validity rules can be applied.

verify_producer_signature

Verify block proposer signature, using the proposal transaction signature as signing key over blocks header hash.

verify_producer_transaction

Verify WASM execution, signatures, and ZK proofs for a given producer Transaction, and apply it to the provided overlay. Returns transaction signature public key.

verify_proposal

Verify given Proposal against provided consensus state

verify_transaction

Verify WASM execution, signatures, and ZK proofs for a given Transaction, and apply it to the provided overlay.

verify_transactions

Verify a set of Transaction in sequence and apply them if all are valid. In case any of the transactions fail, they will be returned to the caller. The function takes a boolean called write which tells it to actually write the state transitions to the database.