Function darkfi::validator::verification::verify_transaction

pub async fn verify_transaction(
    overlay: &BlockchainOverlayPtr,
    verifying_block_height: u32,
    block_target: u32,
    tx: &Transaction,
    tree: &mut MerkleTree,
    verifying_keys: &mut HashMap<[u8; 32], HashMap<String, VerifyingKey>>,
    verify_fee: bool,
) -> Result<GasData>

Verify WASM execution, signatures, and ZK proofs for a given Transaction, and apply it to the provided overlay. Additionally, append its hash to the provided Merkle tree.