darkfi

Module zk

Re-exports

pub use vm::ZkCircuit;

pub use vm_heap::empty_witnesses;

pub use vm_heap::Witness;

pub use proof::Proof;

pub use proof::ProvingKey;

pub use proof::VerifyingKey;

Modules

debug 🔒

gadget

ZK gadget implementations

halo2

proof

Proof creation API

tracer 🔒

Trace computation of intermediate values in circuit

vm

Halo2 zkas virtual machine

vm_heap

VM heap variable definitions and utility functions VM heap type abstractions

Enums

DebugOpValue

Functions

assign_free_advice

export_witness_json

Export witness.json which can be used by zkrunner for debugging circuits Note that this function makes liberal use of unwraps so it could panic.

import_witness_json

Import witness.json which can be used to debug or benchmark circuits. Note that if the path or provided json is incorrect then this function will panic.

zkas_type_checks

Call this before Proof::create() to perform type checks on the witnesses and check the amount of provided instances are correct.