Module darkfi::zk::gadget

ZK gadget implementations

Modules

arithmetic

Base field arithmetic gadget

cond_select

Conditional selection

is_equal

is_equal comparison gadget

is_zero

is_zero comparison gadget

less_than

Field-native less than comparison gadget with a lookup table Given two values: - a, a NUM_OF_BITS-length value and - b, an arbitrary field element, this gadget constrains them in the following way: - in strict mode, a is constrained to be strictly less than b; - else, a is constrained to be less than or equal to b.

native_range_check

Field-native range check gadget with a lookup table

small_range_check

Small range check, 0..8 bits

smt

Poseidon-based sparse Merkle tree chip

zero_cond

Conditional selection based on lhs (will output lhs if lhs==0, otherwise rhs)