darkfi::zk::halo2

Module dev

Tools for developing circuits.

Modules

cost

Developer tools for investigating the cost of a circuit.

metadata

Metadata about circuits.

Structs

CircuitCost

Measures a circuit to determine its costs, and explain what contributes to them.

CircuitGates

A struct for collecting and displaying the gates within a circuit.

CircuitLayout

Graphical renderer for circuit layouts.

MockProver

A test prover for debugging circuits.

TracingFloorPlanner

A helper type that augments a FloorPlanner with [tracing] spans and events.

Enums

FailureLocation

The location within the circuit at which a particular VerifyFailure occurred.

VerifyFailure

The reasons why a particular circuit is not satisfied.

Functions

circuit_dot_graph

Builds a dot graph string representing the given circuit.