darkfi::zk::halo2::plonk
Function create_proof

pub fn create_proof<C, E, R, T, ConcreteCircuit>(
    params: &Params<C>,
    pk: &ProvingKey<C>,
    circuits: &[ConcreteCircuit],
    instances: &[&[&[<C as PrimeCurveAffine>::Scalar]]],
    rng: R,
    transcript: &mut T,
) -> Result<(), Error>
where
    C: CurveAffine,
    E: EncodedChallenge<C>,
    R: RngCore,
    T: TranscriptWrite<C, E>,
    ConcreteCircuit: Circuit<<C as PrimeCurveAffine>::Scalar>,

This creates a proof for the provided circuit when given the public parameters params and the proving key ProvingKey that was generated previously for the same circuit. The provided instances are zero-padded internally.