darkfi::zk::halo2::plonk

Trait VerificationStrategy

pub trait VerificationStrategy<'params, C>
where
    C: CurveAffine,
{
    type Output;

    // Required method
    fn process<E>(
        self,
        f: impl FnOnce(MSM<'params, C>) -> Result<Guard<'params, C, E>, Error>,
    ) -> Result<Self::Output, Error>
       where E: EncodedChallenge<C>;
}

Trait representing a strategy for verifying Halo 2 proofs.

Required Associated Types

type Output

The output type of this verification strategy after processing a proof.

Required Methods

fn process<E>( self, f: impl FnOnce(MSM<'params, C>) -> Result<Guard<'params, C, E>, Error>, ) -> Result<Self::Output, Error>
where E: EncodedChallenge<C>,

Obtains an MSM from the verifier strategy and yields back the strategy’s output.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.