darkfi::zk

Module vm_heap

VM heap variable definitions and utility functions VM heap type abstractions

Macros

impl_try_from 🔒

Enums

HeapVar

These represent the witness types inside the circuit

Witness

These represent the witness types outside of the circuit

Functions

empty_witnesses

Helper function for verifiers to generate empty witnesses for a given decoded zkas binary