Module darkfi::zkas

src/zkas is the library holding the zkas toolchain, consisting of a lexer, parser, static/semantic analyzers, a binary compiler, and a binary decoder.

Re-exports

pub use opcode::Opcode;

pub use types::LitType;

pub use types::VarType;

pub use lexer::Lexer;

pub use parser::Parser;

pub use analyzer::Analyzer;

pub use compiler::Compiler;

pub use decoder::ZkBinary;

Modules

analyzer

Analyzer module

ast

Language AST

compiler

Compiler module

decoder

Decoder module

error 🔒

Error emitter

lexer

Lexer module

opcode

Language opcodes

parser

Parser module

types

Language types